fix: address review on the @Observable migration
- BroadcasterEventModifier now matches the .onReceive semantics it documents:
  the action is refreshed on every body evaluation via a @State-held box, and
  a change of broadcaster identity resubscribes. All four current call sites
  were safe, but the trap was invisible in public package API.
- Move the SwiftUI onEvent glue out of Persistence into AppShared. Broadcaster
  itself stays in Persistence (it backs GRDB table observation); DocumentList
  no longer imports Persistence purely for a view modifier.
- Drop the unnecessary nonisolated(unsafe) on the two @ObservationIgnored task
  properties. @Observable does not require it: @ObservationIgnored leaves them
  plain stored properties, which a nonisolated deinit may read.
- Trim the unused Subscription.store(in:) overloads and the Hashable
  conformance that only existed to serve them. sink(_:) stays — it is what
  onEvent is built on.
- ConnectionManager is final (no subclasses; avoids the subclass-observation
  footgun) and Subscription is plainly Sendable (its only stored property is
  an immutable Task, which is already Sendable).
