+++ Persistence/Sources/Persistence/Observation/Broadcaster.swift
@@ -12,11 +12,11 @@ import Foundation
 /// `DocumentStore` / `ConnectionManager` (Stage 6) and will back the
 /// `CacheChange` signal in later offline-cache stages.
 ///
-/// For SwiftUI consumers, prefer ``sink(_:)`` (returns a ``Subscription``
-/// that can be stored in `@State` via `.store(in:)`) over a bare
-/// `.task { for await … in subscribe() }` — `.task` is cancelled when the
-/// attached view disappears (e.g. on NavigationStack push), which drops
-/// events while a pushed child view is on screen.
+/// For SwiftUI consumers, prefer AppShared's `.onEvent(from:perform:)` — or
+/// ``sink(_:)`` directly, holding the returned ``Subscription`` in `@State` —
+/// over a bare `.task { for await … in subscribe() }`: `.task` is cancelled
+/// when the attached view disappears (e.g. on NavigationStack push), which
+/// drops events while a pushed child view is on screen.
 public final class Broadcaster<Element: Sendable>: @unchecked Sendable {
   private let lock = NSLock()
   private var continuations: [UUID: AsyncStream<Element>.Continuation] = [:]
