+++ swift-paperless/Views/Document/DocumentList.swift
@@ -10,7 +10,6 @@ import DataModel
 import Foundation
 import Networking
 import Nuke
-import Persistence
 import SwiftUI
 import os
